place the tests inside Mnemonic.js and add new test
